Update OneClickArchiver.js
Massive change.  All page parameter detection items are now at the top of the page and run before the page is polluted with useless `|Archive` links.  Next steps are to work on the ability of the script to put the header on pages when it is creating them and to respect "maxpagesize".